requirements.txt
Flask==3.1.0
pymongo==4.10.1
pandas==2.2.3
flask-cors==5.0.0
python-dotenv==1.0.1